+++ src/data/sample-presets.ts
@@ -1,0 +1,430 @@
+import type { FloorPlanData } from "@/types/floorplan";
+
+export type SamplePreset = { name: string; data: FloorPlanData };
+
+export const SAMPLE_PRESETS: SamplePreset[] = [
+    {
+        name: "Studio Apartment",
+        data: {
+            grid: [
+                [false, false, false, false, false, false, false, false, false, false],
+                [false, true, true, true, true, true, true, true, true, false],
+                [false, true, true, true, true, true, true, true, true, false],
+                [false, true, true, true, false, true, true, true, true, false],
+                [false, true, true, true, false, true, true, true, true, false],
+                [false, true, true, true, false, false, true, true, true, false],
+                [false, true, true, true, true, true, true, true, true, false],
+                [false, false, false, false, false, false, false, false, false, false],
+            ],
+            rooms: [
+                { type: "living_room", label: "Living Area", x: 1, y: 1, w: 3, h: 5 },
+                { type: "bedroom", label: "Sleeping Nook", x: 5, y: 1, w: 4, h: 3 },
+                { type: "kitchen", label: "Kitchen", x: 6, y: 5, w: 3, h: 2 },
+                { type: "bathroom", label: "Bathroom", x: 5, y: 5, w: 1, h: 1 },
+            ],
+        },
+    },
+    {
+        name: "One-Bedroom",
+        data: {
+            grid: [
+                [
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                ],
+                [
+                    false,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    false,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    false,
+                ],
+                [
+                    false,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    false,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    false,
+                ],
+                [
+                    false,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    false,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    false,
+                ],
+                [
+                    false,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    false,
+                ],
+                [
+                    false,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    false,
+                ],
+                [
+                    false,
+                    false,
+                    false,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                ],
+                [
+                    false,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    false,
+                    true,
+                    true,
+                    true,
+                    true,
+                    false,
+                ],
+                [
+                    false,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    false,
+                    true,
+                    true,
+                    true,
+                    true,
+                    false,
+                ],
+                [
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                ],
+            ],
+            rooms: [
+                { type: "living_room", label: "Living Room", x: 1, y: 1, w: 5, h: 5 },
+                { type: "bedroom", label: "Bedroom", x: 7, y: 1, w: 6, h: 5 },
+                { type: "hallway", label: "Hall", x: 3, y: 6, w: 5, h: 1 },
+                { type: "kitchen", label: "Kitchen", x: 1, y: 7, w: 7, h: 2 },
+                { type: "bathroom", label: "Bathroom", x: 9, y: 7, w: 4, h: 2 },
+            ],
+        },
+    },
+    {
+        name: "L-Shaped House",
+        data: {
+            grid: [
+                [
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                ],
+                [
+                    false,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    false,
+                    true,
+                    true,
+                    true,
+                    true,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                ],
+                [
+                    false,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    false,
+                    true,
+                    true,
+                    true,
+                    true,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                ],
+                [
+                    false,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    false,
+                    true,
+                    true,
+                    true,
+                    true,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                ],
+                [
+                    false,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                ],
+                [
+                    false,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                ],
+                [
+                    false,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                ],
+                [
+                    false,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    false,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    false,
+                ],
+                [
+                    false,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    false,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    false,
+                ],
+                [
+                    false,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    false,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    false,
+                ],
+                [
+                    false,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    true,
+                    false,
+                ],
+                [
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                    false,
+                ],
+            ],
+            rooms: [
+                { type: "living_room", label: "Living Room", x: 1, y: 1, w: 5, h: 5 },
+                { type: "bedroom", label: "Bedroom", x: 7, y: 1, w: 4, h: 5 },
+                { type: "kitchen", label: "Kitchen", x: 1, y: 6, w: 5, h: 5 },
+                { type: "office", label: "Office", x: 7, y: 7, w: 8, h: 4 },
+            ],
+        },
+    },
+];
